+++ tests/testthat/test-ddml_ate.R
@@ -157,7 +157,8 @@ test_that("summary.ddml_ate computes with a single model", {
   inf_res <- summary(ddml_ate_fit)
   capture_output({print(inf_res)}, print = FALSE)
   # Check output with expectations
-  expect_equal(length(inf_res), 4)
+  expect_s3_class(inf_res, "summary.ddml")
+  expect_equal(dim(inf_res$inf_results), c(1, 4, 1))
 })#TEST_THAT
 
 test_that("summary.ddml_ate computes with a single model and dependence", {
@@ -186,7 +187,8 @@ test_that("summary.ddml_ate computes with a single model and dependence", {
   inf_res <- summary(ddml_ate_fit)
   capture_output({print(inf_res)}, print = FALSE)
   # Check output with expectations
-  expect_equal(length(inf_res), 4)
+  expect_s3_class(inf_res, "summary.ddml")
+  expect_equal(dim(inf_res$inf_results), c(1, 4, 1))
 })#TEST_THAT
 
 test_that("summary.ddml_ate computes with multiple ensemble procedures", {
@@ -212,5 +214,6 @@ test_that("summary.ddml_ate computes with multiple ensemble procedures", {
   inf_res <- summary(ddml_ate_fit)
   capture_output({print(inf_res)}, print = FALSE)
   # Check output with expectations
-  expect_equal(length(inf_res), 16)
+  expect_s3_class(inf_res, "summary.ddml")
+  expect_equal(dim(inf_res$inf_results), c(1, 4, 4))
 })#TEST_THAT